Resolve circular dependencies between csscasc.ts and selectors.ts
